pyproject.toml
[project]
name = "agent-social-network"
version = "0.1.0"
description = "P2P Social Network of AI Agents — private capability exchange"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "mcp>=1.25",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "openai>=1.50",
    "pydantic-settings>=2.5",
    "httpx>=0.27",
    "structlog>=24.0",
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    "aiosqlite>=0.20",
    "pynacl>=1.5",
    "kademlia>=2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.34",
    "ruff>=0.8",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: tests that load ML models (sentence-transformers)",
    "integration: full server integration tests",
]
